requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Pillow==10.1.0
pillow-heif==0.13.1
reportlab==4.0.7
python-docx==1.1.0
pypandoc==1.12
werkzeug==3.0.1